// app/types/project.ts
export interface Project {
  budgetGrade: string;
  classification: string;
  clientName: string;
  createdAt: string;
  description: string;
  drawingCompletionDate: string;
  endDate: string;
  equipmentCategory: string;
  equipmentNumbers: string;
  installationDate: string;
  isProject: string;
  name: string;
  notes: string;
  projectEndDate: string;
  projectNumber: string;
  projectStartDate: string;
  spare1: string;
  spare2: string;
  spare3: string;
  startDate: string;
  status: string;
  updatedAt: string;
} 